requirements.txt
# ͨ��ģ������
pyserial==3.5
pymodbus==3.0.2
paho-mqtt==1.6.1 
# ��˿������
fastapi==0.104.1
uvicorn==0.24.0.post1

# �ƶ�Эͬ����
paho-mqtt==1.6.1

# ��Ե��������
numpy==1.26.2
scikit-learn==1.3.2

# ���߿�
python-dotenv==1.0.0
